Split automation integration tests from workspace model
