+++ Tests/QuillCodeAppTests/WorkspaceModelTests.swift
@@ -2948,674 +2948,6 @@ final class WorkspaceModelTests: XCTestCase {
         XCTAssertEqual(try projectStore.load().map(\.name), ["QuillCode"])
     }
 
-    func testModelPersistsAutomationChanges() throws {
-        let root = try makeTempDirectory()
-        let paths = QuillCodePaths(home: root.appendingPathComponent(".quillcode"))
-        try paths.ensure()
-        let automationStore = JSONAutomationStore(fileURL: paths.automationsFile)
-        let model = QuillCodeWorkspaceModel(automationStore: automationStore)
-
-        model.setAutomations([
-            QuillAutomation(
-                title: "Morning check",
-                detail: "Summarize the repo state.",
-                kind: .workspaceSchedule,
-                scheduleKind: .cron,
-                scheduleDescription: "Every morning"
-            )
-        ])
-
-        XCTAssertEqual(try automationStore.load().map(\.title), ["Morning check"])
-    }
-
-    func testAutomationCommandsCreatePauseResumeAndDeletePersistedFollowUp() throws {
-        let root = try makeTempDirectory()
-        let paths = QuillCodePaths(home: root.appendingPathComponent(".quillcode"))
-        try paths.ensure()
-        let automationStore = JSONAutomationStore(fileURL: paths.automationsFile)
-        let thread = ChatThread(title: "Launch plan")
-        let model = QuillCodeWorkspaceModel(
-            root: QuillCodeRootState(threads: [thread], selectedThreadID: thread.id),
-            automationStore: automationStore
-        )
-
-        XCTAssertTrue(model.runWorkspaceCommand("automation-create-thread-follow-up", workspaceRoot: root))
-
-        let created = try XCTUnwrap(try automationStore.load().first)
-        XCTAssertEqual(created.title, "Follow up: Launch plan")
-        XCTAssertEqual(created.threadID, thread.id)
-        XCTAssertEqual(created.status, .active)
-        XCTAssertEqual(model.surface().automations.statusLabel, "1 active")
-
-        XCTAssertTrue(model.runWorkspaceCommand("automation-pause:\(created.id.uuidString)", workspaceRoot: root))
-        XCTAssertEqual(try automationStore.load().first?.status, .paused)
-        XCTAssertEqual(model.surface().automations.workflows.first?.primaryActionTitle, "Resume")
-
-        XCTAssertTrue(model.runWorkspaceCommand("automation-resume:\(created.id.uuidString)", workspaceRoot: root))
-        XCTAssertEqual(try automationStore.load().first?.status, .active)
-        XCTAssertEqual(model.surface().automations.workflows.first?.primaryActionTitle, "Pause")
-
-        XCTAssertTrue(model.runWorkspaceCommand("automation-delete:\(created.id.uuidString)", workspaceRoot: root))
-        XCTAssertEqual(try automationStore.load(), [])
-        XCTAssertEqual(model.surface().automations.workflows.map(\.title), [
-            "Thread follow-ups",
-            "Workspace schedules",
-            "Monitors"
-        ])
-    }
-
-    func testScheduledThreadFollowUpsPersistConcreteRunTimes() throws {
-        let root = try makeTempDirectory()
-        let paths = QuillCodePaths(home: root.appendingPathComponent(".quillcode"))
-        try paths.ensure()
-        let automationStore = JSONAutomationStore(fileURL: paths.automationsFile)
-        let thread = ChatThread(title: "Launch plan")
-        let model = QuillCodeWorkspaceModel(
-            root: QuillCodeRootState(threads: [thread], selectedThreadID: thread.id),
-            automationStore: automationStore
-        )
-        let now = Date(timeIntervalSince1970: 1_000)
-        var calendar = Calendar(identifier: .gregorian)
-        calendar.timeZone = TimeZone(secondsFromGMT: 0)!
-
-        let tenMinute = try XCTUnwrap(model.createThreadFollowUpAutomation(after: 600, now: now))
-        let tomorrow = try XCTUnwrap(model.createTomorrowMorningThreadFollowUpAutomation(
-            now: now,
-            calendar: calendar
-        ))
-
-        XCTAssertEqual(tenMinute.scheduleDescription, "In 10 minutes")
-        XCTAssertEqual(tenMinute.nextRunAt, now.addingTimeInterval(600))
-        XCTAssertEqual(tomorrow.scheduleDescription, "Tomorrow at 9:00 AM")
-        XCTAssertEqual(
-            tomorrow.nextRunAt,
-            calendar.date(from: DateComponents(
-                calendar: calendar,
-                timeZone: calendar.timeZone,
-                year: 1970,
-                month: 1,
-                day: 2,
-                hour: 9,
-                minute: 0,
-                second: 0
-            ))
-        )
-
-        let saved = try automationStore.load()
-        XCTAssertEqual(saved.map(\.scheduleDescription), ["In 10 minutes", "Tomorrow at 9:00 AM"])
-        XCTAssertEqual(saved.map(\.threadID), [thread.id, thread.id])
-    }
-
-    func testNaturalLanguageScheduledThreadFollowUpsPersistConcreteRunTimes() throws {
-        let root = try makeTempDirectory()
-        let paths = QuillCodePaths(home: root.appendingPathComponent(".quillcode"))
-        try paths.ensure()
-        let automationStore = JSONAutomationStore(fileURL: paths.automationsFile)
-        let thread = ChatThread(title: "Launch plan")
-        let model = QuillCodeWorkspaceModel(
-            root: QuillCodeRootState(threads: [thread], selectedThreadID: thread.id),
-            automationStore: automationStore
-        )
-        let now = Date(timeIntervalSince1970: 1_000)
-        var calendar = Calendar(identifier: .gregorian)
-        calendar.timeZone = TimeZone(secondsFromGMT: 0)!
-
-        let relative = try XCTUnwrap(model.createThreadFollowUpAutomation(
-            matching: "in 45 minutes",
-            now: now,
-            calendar: calendar
-        ))
-        let tomorrow = try XCTUnwrap(model.createThreadFollowUpAutomation(
-            matching: "tomorrow at 9:30 PM",
-            now: now,
-            calendar: calendar
-        ))
-
-        XCTAssertEqual(relative.scheduleDescription, "In 45 minutes")
-        XCTAssertEqual(relative.nextRunAt, now.addingTimeInterval(45 * 60))
-        XCTAssertEqual(tomorrow.scheduleDescription, "Tomorrow at 9:30 PM")
-        XCTAssertEqual(
-            tomorrow.nextRunAt,
-            calendar.date(from: DateComponents(
-                calendar: calendar,
-                timeZone: calendar.timeZone,
-                year: 1970,
-                month: 1,
-                day: 2,
-                hour: 21,
-                minute: 30,
-                second: 0
-            ))
-        )
-
-        let saved = try automationStore.load()
-        XCTAssertEqual(saved.map(\.scheduleDescription), ["In 45 minutes", "Tomorrow at 9:30 PM"])
-        XCTAssertEqual(saved.map(\.threadID), [thread.id, thread.id])
-    }
-
-    func testScheduledWorkspaceChecksPersistConcreteRunTimes() throws {
-        let root = try makeTempDirectory()
-        let paths = QuillCodePaths(home: root.appendingPathComponent(".quillcode"))
-        try paths.ensure()
-        let automationStore = JSONAutomationStore(fileURL: paths.automationsFile)
-        let project = ProjectRef(name: "QuillCode", path: root.path)
-        let model = QuillCodeWorkspaceModel(
-            root: QuillCodeRootState(projects: [project], selectedProjectID: project.id),
-            automationStore: automationStore
-        )
-        let now = Date(timeIntervalSince1970: 1_000)
-        var calendar = Calendar(identifier: .gregorian)
-        calendar.timeZone = TimeZone(secondsFromGMT: 0)!
-
-        let tenMinute = try XCTUnwrap(model.createWorkspaceScheduleAutomation(after: 600, now: now))
-        let tomorrow = try XCTUnwrap(model.createTomorrowMorningWorkspaceScheduleAutomation(
-            now: now,
-            calendar: calendar
-        ))
-
-        XCTAssertEqual(tenMinute.title, "Workspace check: QuillCode")
-        XCTAssertEqual(tenMinute.scheduleDescription, "In 10 minutes")
-        XCTAssertEqual(tenMinute.nextRunAt, now.addingTimeInterval(600))
-        XCTAssertEqual(tomorrow.scheduleDescription, "Tomorrow at 9:00 AM")
-        XCTAssertEqual(
-            tomorrow.nextRunAt,
-            calendar.date(from: DateComponents(
-                calendar: calendar,
-                timeZone: calendar.timeZone,
-                year: 1970,
-                month: 1,
-                day: 2,
-                hour: 9,
-                minute: 0,
-                second: 0
-            ))
-        )
-
-        let saved = try automationStore.load()
-        XCTAssertEqual(saved.map(\.kind), [.workspaceSchedule, .workspaceSchedule])
-        XCTAssertEqual(saved.map(\.scheduleDescription), ["In 10 minutes", "Tomorrow at 9:00 AM"])
-        XCTAssertEqual(saved.map(\.projectID), [project.id, project.id])
-        XCTAssertEqual(saved.map(\.threadID), [nil, nil])
-    }
-
-    func testNaturalLanguageScheduledWorkspaceChecksPersistConcreteRunTimes() throws {
-        let root = try makeTempDirectory()
-        let paths = QuillCodePaths(home: root.appendingPathComponent(".quillcode"))
-        try paths.ensure()
-        let automationStore = JSONAutomationStore(fileURL: paths.automationsFile)
-        let project = ProjectRef(name: "QuillCode", path: root.path)
-        let model = QuillCodeWorkspaceModel(
-            root: QuillCodeRootState(projects: [project], selectedProjectID: project.id),
-            automationStore: automationStore
-        )
-        let now = Date(timeIntervalSince1970: 1_000)
-        var calendar = Calendar(identifier: .gregorian)
-        calendar.timeZone = TimeZone(secondsFromGMT: 0)!
-
-        let relative = try XCTUnwrap(model.createWorkspaceScheduleAutomation(
-            matching: "in 2 hours",
-            now: now,
-            calendar: calendar
-        ))
-        let tomorrow = try XCTUnwrap(model.createWorkspaceScheduleAutomation(
-            matching: "tomorrow at 8:15 AM",
-            now: now,
-            calendar: calendar
-        ))
-
-        XCTAssertEqual(relative.scheduleDescription, "In 2 hours")
-        XCTAssertEqual(relative.nextRunAt, now.addingTimeInterval(2 * 60 * 60))
-        XCTAssertEqual(tomorrow.scheduleDescription, "Tomorrow at 8:15 AM")
-        XCTAssertEqual(
-            tomorrow.nextRunAt,
-            calendar.date(from: DateComponents(
-                calendar: calendar,
-                timeZone: calendar.timeZone,
-                year: 1970,
-                month: 1,
-                day: 2,
-                hour: 8,
-                minute: 15,
-                second: 0
-            ))
-        )
-
-        let saved = try automationStore.load()
-        XCTAssertEqual(saved.map(\.kind), [.workspaceSchedule, .workspaceSchedule])
-        XCTAssertEqual(saved.map(\.scheduleDescription), ["In 2 hours", "Tomorrow at 8:15 AM"])
-        XCTAssertEqual(saved.map(\.projectID), [project.id, project.id])
-    }
-
-    func testNaturalLanguageRecurringWorkspaceChecksPersistRecurrence() throws {
-        let root = try makeTempDirectory()
-        let paths = QuillCodePaths(home: root.appendingPathComponent(".quillcode"))
-        try paths.ensure()
-        let automationStore = JSONAutomationStore(fileURL: paths.automationsFile)
-        let project = ProjectRef(name: "QuillCode", path: root.path)
-        let model = QuillCodeWorkspaceModel(
-            root: QuillCodeRootState(projects: [project], selectedProjectID: project.id),
-            automationStore: automationStore
-        )
-        let now = Date(timeIntervalSince1970: 1_000)
-
-        let recurring = try XCTUnwrap(model.createWorkspaceScheduleAutomation(
-            matching: "every 2 hours",
-            now: now
-        ))
-
-        XCTAssertEqual(recurring.scheduleDescription, "Every 2 hours")
-        XCTAssertEqual(recurring.recurrence, QuillAutomationRecurrence(interval: 2, unit: .hours))
-        XCTAssertEqual(recurring.nextRunAt, now.addingTimeInterval(2 * 60 * 60))
-
-        let saved = try XCTUnwrap(try automationStore.load().first)
-        XCTAssertEqual(saved.scheduleDescription, "Every 2 hours")
-        XCTAssertEqual(saved.recurrence, QuillAutomationRecurrence(interval: 2, unit: .hours))
-        XCTAssertEqual(saved.nextRunAt, now.addingTimeInterval(2 * 60 * 60))
-    }
-
-    func testSlashFollowUpSchedulesCurrentThread() async throws {
-        let root = try makeTempDirectory()
-        let paths = QuillCodePaths(home: root.appendingPathComponent(".quillcode"))
-        try paths.ensure()
-        let automationStore = JSONAutomationStore(fileURL: paths.automationsFile)
-        let thread = ChatThread(title: "Launch plan")
-        let model = QuillCodeWorkspaceModel(
-            root: QuillCodeRootState(threads: [thread], selectedThreadID: thread.id),
-            automationStore: automationStore
-        )
-
-        model.setDraft("/follow-up in 45 minutes")
-        await model.submitComposer(workspaceRoot: root)
-
-        let saved = try automationStore.load()
-        XCTAssertEqual(saved.count, 1)
-        XCTAssertEqual(saved.first?.title, "Follow up: Launch plan")
-        XCTAssertEqual(saved.first?.scheduleDescription, "In 45 minutes")
-        XCTAssertNotNil(saved.first?.nextRunAt)
-        XCTAssertEqual(model.selectedThread?.messages.last?.content, "Scheduled a thread follow-up for In 45 minutes.")
-        XCTAssertEqual(model.surface().automations.statusLabel, "1 active")
-    }
-
-    func testSlashWorkspaceCheckSchedulesSelectedProject() async throws {
-        let root = try makeTempDirectory()
-        let paths = QuillCodePaths(home: root.appendingPathComponent(".quillcode"))
-        try paths.ensure()
-        let automationStore = JSONAutomationStore(fileURL: paths.automationsFile)
-        let project = ProjectRef(name: "QuillCode", path: root.path)
-        let model = QuillCodeWorkspaceModel(
-            root: QuillCodeRootState(projects: [project], selectedProjectID: project.id),
-            automationStore: automationStore
-        )
-
-        model.setDraft("/workspace-check tomorrow at 8:15 AM")
-        await model.submitComposer(workspaceRoot: root)
-
-        let saved = try automationStore.load()
-        XCTAssertEqual(saved.count, 1)
-        XCTAssertEqual(saved.first?.title, "Workspace check: QuillCode")
-        XCTAssertEqual(saved.first?.projectID, project.id)
-        XCTAssertEqual(saved.first?.kind, .workspaceSchedule)
-        XCTAssertEqual(saved.first?.scheduleDescription, "Tomorrow at 8:15 AM")
-        XCTAssertNotNil(saved.first?.nextRunAt)
-        XCTAssertEqual(model.selectedThread?.messages.last?.content, "Scheduled a workspace check for Tomorrow at 8:15 AM.")
-        XCTAssertEqual(model.surface().automations.statusLabel, "1 active")
-    }
-
-    func testSlashWorkspaceCheckSchedulesRecurringProjectAutomation() async throws {
-        let root = try makeTempDirectory()
-        let paths = QuillCodePaths(home: root.appendingPathComponent(".quillcode"))
-        try paths.ensure()
-        let automationStore = JSONAutomationStore(fileURL: paths.automationsFile)
-        let project = ProjectRef(name: "QuillCode", path: root.path)
-        let model = QuillCodeWorkspaceModel(
-            root: QuillCodeRootState(projects: [project], selectedProjectID: project.id),
-            automationStore: automationStore
-        )
-
-        model.setDraft("/workspace-check daily")
-        await model.submitComposer(workspaceRoot: root)
-
-        let saved = try XCTUnwrap(try automationStore.load().first)
-        XCTAssertEqual(saved.title, "Workspace check: QuillCode")
-        XCTAssertEqual(saved.projectID, project.id)
-        XCTAssertEqual(saved.kind, .workspaceSchedule)
-        XCTAssertEqual(saved.scheduleDescription, "Every day")
-        XCTAssertEqual(saved.recurrence, QuillAutomationRecurrence(interval: 1, unit: .days))
-        XCTAssertNotNil(saved.nextRunAt)
-        XCTAssertEqual(model.selectedThread?.messages.last?.content, "Scheduled a workspace check for Every day.")
-        XCTAssertEqual(model.surface().automations.statusLabel, "1 active")
-    }
-
-    func testAutomationRunCreatesFollowUpThreadAndPersistsRunMetadata() throws {
-        let root = try makeTempDirectory()
-        let paths = QuillCodePaths(home: root.appendingPathComponent(".quillcode"))
-        try paths.ensure()
-        let automationStore = JSONAutomationStore(fileURL: paths.automationsFile)
-        let threadStore = JSONThreadStore(directory: paths.threadsDirectory)
-        let source = ChatThread(title: "Launch plan", messages: [
-            .init(role: .user, content: "latest question"),
-            .init(role: .tool, content: #"{"internal":"skip"}"#),
-            .init(role: .assistant, content: "latest answer")
-        ])
-        let automation = QuillAutomation(
-            title: "Follow up: Launch plan",
-            detail: "Resume this thread with the same project, model, and context.",
-            kind: .threadFollowUp,
-            scheduleKind: .heartbeat,
-            scheduleDescription: "Manual follow-up",
-            threadID: source.id,
-            nextRunAt: Date(timeIntervalSince1970: 1)
-        )
-        let model = QuillCodeWorkspaceModel(
-            root: QuillCodeRootState(threads: [source], selectedThreadID: source.id),
-            threadStore: threadStore,
-            automationStore: automationStore
-        )
-        model.setAutomations([automation])
-
-        XCTAssertTrue(model.runWorkspaceCommand("automation-run:\(automation.id.uuidString)", workspaceRoot: root))
-
-        let followUpID = try XCTUnwrap(model.root.selectedThreadID)
-        XCTAssertNotEqual(followUpID, source.id)
-        let followUp = try XCTUnwrap(model.root.threads.first { $0.id == followUpID })
-        XCTAssertEqual(followUp.title, "Follow-up: Launch plan")
-        XCTAssertEqual(followUp.messages.map(\.content), ["latest question", "latest answer"])
-        XCTAssertFalse(followUp.messages.contains { $0.role == .tool })
-        XCTAssertEqual(followUp.events.first?.summary, "Automation ran: Follow up: Launch plan")
-        XCTAssertEqual(followUp.events.first?.payloadJSON, automation.id.uuidString)
-
-        let savedThread = try threadStore.load(followUpID)
-        XCTAssertEqual(savedThread.title, "Follow-up: Launch plan")
-        let savedAutomation = try XCTUnwrap(try automationStore.load().first)
-        XCTAssertNotNil(savedAutomation.lastRunAt)
-        XCTAssertNil(savedAutomation.nextRunAt)
-        XCTAssertEqual(model.surface().automations.workflows.first?.statusLabel, "Ran")
-        XCTAssertEqual(model.surface().automations.statusLabel, "1 active")
-    }
-
-    func testCreateWorkspaceScheduleCommandPersistsSelectedProjectAutomation() throws {
-        let root = try makeTempDirectory()
-        let paths = QuillCodePaths(home: root.appendingPathComponent(".quillcode"))
-        try paths.ensure()
-        let automationStore = JSONAutomationStore(fileURL: paths.automationsFile)
-        let project = ProjectRef(name: "QuillCode", path: root.path)
-        let model = QuillCodeWorkspaceModel(
-            root: QuillCodeRootState(projects: [project], selectedProjectID: project.id),
-            automationStore: automationStore
-        )
-
-        XCTAssertTrue(model.runWorkspaceCommand("automation-create-workspace-schedule", workspaceRoot: root))
-
-        let automation = try XCTUnwrap(model.automations.items.first)
-        XCTAssertEqual(automation.title, "Workspace check: QuillCode")
-        XCTAssertEqual(automation.detail, "Create a scheduled workspace-check thread for QuillCode.")
-        XCTAssertEqual(automation.kind, .workspaceSchedule)
-        XCTAssertEqual(automation.scheduleKind, .cron)
-        XCTAssertEqual(automation.scheduleDescription, "Manual workspace check")
-        XCTAssertEqual(automation.projectID, project.id)
-        XCTAssertNil(automation.threadID)
-        XCTAssertNil(automation.nextRunAt)
-        XCTAssertTrue(model.surface().automations.isVisible)
-
-        let saved = try XCTUnwrap(try automationStore.load().first)
-        XCTAssertEqual(saved.id, automation.id)
-        XCTAssertEqual(saved.projectID, project.id)
-        XCTAssertEqual(saved.kind, .workspaceSchedule)
-    }
-
-    func testRunWorkspaceScheduleCreatesScheduledProjectThread() throws {
-        let root = try makeTempDirectory()
-        let paths = QuillCodePaths(home: root.appendingPathComponent(".quillcode"))
-        try paths.ensure()
-        let threadStore = JSONThreadStore(directory: paths.threadsDirectory)
-        let automationStore = JSONAutomationStore(fileURL: paths.automationsFile)
-        let project = ProjectRef(name: "QuillCode", path: root.path)
-        let automation = QuillAutomation(
-            title: "Workspace check: QuillCode",
-            detail: "Create a scheduled workspace-check thread for QuillCode.",
-            kind: .workspaceSchedule,
-            scheduleKind: .cron,
-            scheduleDescription: "Manual workspace check",
-            projectID: project.id
-        )
-        let model = QuillCodeWorkspaceModel(
-            root: QuillCodeRootState(projects: [project], selectedProjectID: project.id),
-            threadStore: threadStore,
-            automationStore: automationStore
-        )
-        model.setAutomations([automation])
-
-        XCTAssertTrue(model.runWorkspaceCommand("automation-run:\(automation.id.uuidString)", workspaceRoot: root))
-
-        let scheduledID: UUID = try XCTUnwrap(model.root.selectedThreadID)
-        let scheduled: ChatThread = try XCTUnwrap(model.root.threads.first { $0.id == scheduledID })
-        XCTAssertEqual(scheduled.title, "Scheduled check: QuillCode")
-        XCTAssertEqual(scheduled.projectID, project.id)
-        XCTAssertEqual(scheduled.messages.map(\.content), [
-            "Run the scheduled workspace check for QuillCode. Start with project status, recent changes, local actions, and anything needing attention."
-        ])
-        XCTAssertEqual(scheduled.events.first?.summary, "Automation ran: Workspace check: QuillCode")
-        XCTAssertEqual(scheduled.events.first?.payloadJSON, automation.id.uuidString)
-
-        let savedThread = try threadStore.load(scheduledID)
-        XCTAssertEqual(savedThread.title, "Scheduled check: QuillCode")
-        let savedAutomation = try XCTUnwrap(try automationStore.load().first)
-        XCTAssertNotNil(savedAutomation.lastRunAt)
-        XCTAssertNil(savedAutomation.nextRunAt)
-        XCTAssertEqual(model.surface().automations.workflows.first?.statusLabel, "Ran")
-        XCTAssertEqual(model.surface().automations.statusLabel, "1 active")
-    }
-
-    func testRunDueAutomationsRunsActiveDueThreadAndWorkspaceSchedules() throws {
-        let root = try makeTempDirectory()
-        let paths = QuillCodePaths(home: root.appendingPathComponent(".quillcode"))
-        try paths.ensure()
-        let automationStore = JSONAutomationStore(fileURL: paths.automationsFile)
-        let now = Date(timeIntervalSince1970: 100)
-        let project = ProjectRef(name: "QuillCode", path: root.path)
-        let source = ChatThread(title: "Due follow-up", messages: [
-            .init(role: .user, content: "summarize tomorrow"),
-            .init(role: .assistant, content: "I will follow up.")
-        ])
-        let due = QuillAutomation(
-            title: "Due follow-up",
-            detail: "Resume this thread.",
-            kind: .threadFollowUp,
-            scheduleKind: .heartbeat,
-            scheduleDescription: "Now",
-            threadID: source.id,
-            nextRunAt: Date(timeIntervalSince1970: 90)
-        )
-        let workspaceDue = QuillAutomation(
-            title: "Due workspace check",
-            detail: "Check workspace.",
-            kind: .workspaceSchedule,
-            scheduleKind: .cron,
-            scheduleDescription: "Now",
-            projectID: project.id,
-            nextRunAt: Date(timeIntervalSince1970: 85)
-        )
-        let future = QuillAutomation(
-            title: "Future follow-up",
-            detail: "Resume later.",
-            kind: .threadFollowUp,
-            scheduleKind: .heartbeat,
-            scheduleDescription: "Later",
-            threadID: source.id,
-            nextRunAt: Date(timeIntervalSince1970: 120)
-        )
-        let paused = QuillAutomation(
-            title: "Paused follow-up",
-            detail: "Do not run.",
-            kind: .threadFollowUp,
-            status: .paused,
-            scheduleKind: .heartbeat,
-            scheduleDescription: "Now",
-            threadID: source.id,
-            nextRunAt: Date(timeIntervalSince1970: 80)
-        )
-        let unsupported = QuillAutomation(
-            title: "Due monitor",
-            detail: "Monitor runner pending.",
-            kind: .monitor,
-            scheduleKind: .event,
-            scheduleDescription: "Now",
-            nextRunAt: Date(timeIntervalSince1970: 70)
-        )
-        let model = QuillCodeWorkspaceModel(
-            root: QuillCodeRootState(
-                projects: [project],
-                selectedProjectID: project.id,
-                threads: [source],
-                selectedThreadID: source.id
-            ),
-            automationStore: automationStore
-        )
-        model.setAutomations([future, paused, unsupported, due, workspaceDue])
-
-        let followUpIDs = model.runDueAutomations(now: now)
-
-        XCTAssertEqual(followUpIDs.count, 2)
-        let followUp = try XCTUnwrap(model.root.threads.first { $0.title == "Follow-up: Due follow-up" })
-        XCTAssertEqual(followUp.title, "Follow-up: Due follow-up")
-        let workspaceCheck = try XCTUnwrap(model.root.threads.first { $0.title == "Scheduled check: QuillCode" })
-        XCTAssertEqual(workspaceCheck.projectID, project.id)
-        XCTAssertTrue([followUp.id, workspaceCheck.id].contains(try XCTUnwrap(model.root.selectedThreadID)))
-
-        let savedAutomations = try automationStore.load()
-        let savedDue = try XCTUnwrap(savedAutomations.first { $0.id == due.id })
-        XCTAssertNotNil(savedDue.lastRunAt)
-        XCTAssertNil(savedDue.nextRunAt)
-        let savedWorkspaceDue = try XCTUnwrap(savedAutomations.first { $0.id == workspaceDue.id })
-        XCTAssertNotNil(savedWorkspaceDue.lastRunAt)
-        XCTAssertNil(savedWorkspaceDue.nextRunAt)
-        XCTAssertEqual(savedAutomations.first { $0.id == future.id }?.nextRunAt, future.nextRunAt)
-        XCTAssertEqual(savedAutomations.first { $0.id == paused.id }?.nextRunAt, paused.nextRunAt)
-        XCTAssertEqual(savedAutomations.first { $0.id == unsupported.id }?.nextRunAt, unsupported.nextRunAt)
-    }
-
-    func testDueRecurringWorkspaceScheduleRunsAndAdvancesNextRun() throws {
-        let root = try makeTempDirectory()
-        let paths = QuillCodePaths(home: root.appendingPathComponent(".quillcode"))
-        try paths.ensure()
-        let automationStore = JSONAutomationStore(fileURL: paths.automationsFile)
-        let runAt = Date(timeIntervalSince1970: floor(Date().timeIntervalSince1970) + 60)
-        let project = ProjectRef(name: "QuillCode", path: root.path)
-        let recurrence = QuillAutomationRecurrence(interval: 2, unit: .hours)
-        let recurring = QuillAutomation(
-            title: "Recurring workspace check",
-            detail: "Check workspace.",
-            kind: .workspaceSchedule,
-            scheduleKind: .cron,
-            scheduleDescription: recurrence.scheduleDescription,
-            projectID: project.id,
-            nextRunAt: runAt.addingTimeInterval(-10),
-            recurrence: recurrence
-        )
-        let model = QuillCodeWorkspaceModel(
-            root: QuillCodeRootState(projects: [project], selectedProjectID: project.id),
-            automationStore: automationStore
-        )
-        model.setAutomations([recurring])
-
-        let reports = model.runDueAutomationReports(now: runAt)
-
-        XCTAssertEqual(reports.count, 1)
-        XCTAssertEqual(model.root.threads.filter { $0.title == "Scheduled check: QuillCode" }.count, 1)
-        let saved = try XCTUnwrap(try automationStore.load().first)
-        XCTAssertNotNil(saved.lastRunAt)
-        XCTAssertEqual(
-            try XCTUnwrap(saved.nextRunAt).timeIntervalSince1970,
-            recurrence.nextRun(after: runAt).timeIntervalSince1970,
-            accuracy: 0.001
-        )
-        XCTAssertEqual(saved.recurrence, recurrence)
-        XCTAssertEqual(model.surface().automations.workflows.first?.statusLabel, "Active")
-
-        _ = model.runDueAutomations(now: runAt.addingTimeInterval(60))
-
-        XCTAssertEqual(model.root.threads.filter { $0.title == "Scheduled check: QuillCode" }.count, 1)
-    }
-
-    func testRunDueAutomationReportsDescribeCreatedFollowUps() throws {
-        let root = try makeTempDirectory()
-        let paths = QuillCodePaths(home: root.appendingPathComponent(".quillcode"))
-        try paths.ensure()
-        let automationStore = JSONAutomationStore(fileURL: paths.automationsFile)
-        let now = Date(timeIntervalSince1970: 100)
-        let source = ChatThread(title: "Due follow-up", messages: [
-            .init(role: .user, content: "summarize tomorrow"),
-            .init(role: .assistant, content: "I will follow up.")
-        ])
-        let due = QuillAutomation(
-            title: "Due follow-up",
-            detail: "Resume this thread.",
-            kind: .threadFollowUp,
-            scheduleKind: .heartbeat,
-            scheduleDescription: "Now",
-            threadID: source.id,
-            nextRunAt: Date(timeIntervalSince1970: 90)
-        )
-        let model = QuillCodeWorkspaceModel(
-            root: QuillCodeRootState(threads: [source], selectedThreadID: source.id),
-            automationStore: automationStore
-        )
-        model.setAutomations([due])
-
-        let reports = model.runDueAutomationReports(now: now)
-
-        let report = try XCTUnwrap(reports.first)
-        XCTAssertEqual(reports.count, 1)
-        XCTAssertEqual(report.automationID, due.id)
-        XCTAssertEqual(report.title, "QuillCode follow-up ready")
-        XCTAssertEqual(report.body, "Follow-up: Due follow-up was created from Due follow-up.")
-        XCTAssertEqual(model.root.selectedThreadID, report.followUpThreadID)
-        XCTAssertEqual(model.root.threads.first?.id, report.followUpThreadID)
-        XCTAssertEqual(model.root.threads.first?.title, "Follow-up: Due follow-up")
-
-        let savedDue = try XCTUnwrap(try automationStore.load().first { $0.id == due.id })
-        XCTAssertNotNil(savedDue.lastRunAt)
-        XCTAssertNil(savedDue.nextRunAt)
-    }
-
-    func testRunDueAutomationsHonorsLimit() throws {
-        let now = Date(timeIntervalSince1970: 100)
-        let source = ChatThread(title: "Launch plan", messages: [
-            .init(role: .user, content: "follow up"),
-            .init(role: .assistant, content: "Noted.")
-        ])
-        let first = QuillAutomation(
-            title: "First follow-up",
-            detail: "Resume this thread.",
-            kind: .threadFollowUp,
-            scheduleKind: .heartbeat,
-            scheduleDescription: "Now",
-            threadID: source.id,
-            nextRunAt: Date(timeIntervalSince1970: 80)
-        )
-        let second = QuillAutomation(
-            title: "Second follow-up",
-            detail: "Resume this thread.",
-            kind: .threadFollowUp,
-            scheduleKind: .heartbeat,
-            scheduleDescription: "Now",
-            threadID: source.id,
-            nextRunAt: Date(timeIntervalSince1970: 90)
-        )
-        let model = QuillCodeWorkspaceModel(root: QuillCodeRootState(
-            threads: [source],
-            selectedThreadID: source.id
-        ))
-        model.setAutomations([second, first])
-
-        let followUpIDs = model.runDueAutomations(now: now, limit: 1)
-
-        XCTAssertEqual(followUpIDs.count, 1)
-        XCTAssertEqual(model.automations.items.first { $0.id == first.id }?.lastRunAt != nil, true)
-        XCTAssertNil(model.automations.items.first { $0.id == first.id }?.nextRunAt)
-        XCTAssertEqual(model.automations.items.first { $0.id == second.id }?.nextRunAt, second.nextRunAt)
-    }
-
     func testBootstrapPersistsAndClearsTrustedRouterAPIKey() throws {
         let paths = QuillCodePaths(home: try makeTempDirectory())
         let bootstrap = QuillCodeWorkspaceBootstrap(paths: paths)
