+++ docs/DECISIONS.md
@@ -216,6 +216,7 @@
 - Workspace project extension integration coverage lives in `WorkspaceProjectExtensionIntegrationTests`, not `WorkspaceModelTests`. Extension manifest loading into project/secondary-pane surfaces, update-command execution, metadata refresh, and success/failure transcript notices cross the workspace model and project metadata loader; keep those flows together outside the model monolith while low-level manifest parsing and shell-call shape tests stay focused.
 - Workspace slash-command integration coverage lives in `WorkspaceSlashCommandIntegrationTests`, not `WorkspaceModelTests`. Command-palette slash prefills, core slash dispatch, local environment action slash execution, local model/mode/thread lifecycle commands, context compaction, and status transcript assertions cross composer submission, command dispatch, workspace surfaces, and transcript creation; keep those flows together while pure transcript copy stays in `WorkspaceSlashCommandTranscriptPlannerTests` and remote/automation-specific slash flows stay with their broader owning features until extracted.
 - Workspace local environment integration coverage lives in `WorkspaceLocalEnvironmentIntegrationTests`, not `WorkspaceModelTests`. Local environment action loading into project state, command-palette execution, metadata-backed environment redaction, bounded working directories, and timeouts cross the workspace model, metadata loader, shell-call planner, and tool cards; keep those flows together outside the model monolith while direct loader behavior stays focused. `/env` transcript behavior belongs to `WorkspaceSlashCommandIntegrationTests` with the rest of slash-command routing.
+- Workspace automation integration coverage lives in `WorkspaceAutomationIntegrationTests`, not `WorkspaceModelTests`. Automation persistence, create/pause/resume/delete commands, scheduled and natural-language follow-ups/workspace checks, slash scheduling commands, due automation execution, recurrence advancement, and run reports cross workspace model, stores, surfaces, transcript creation, and thread persistence; keep those flows together while pure automation record and run-draft logic stays in `WorkspaceAutomationEngineTests`.
 - Worktree request values, tool-call construction, and handoff transcript records live outside `WorkspaceModel`. `WorkspaceWorktreeRequests` owns the public create/remove request structs used by dialogs, SwiftUI, and desktop controllers. `WorkspaceWorktreeToolCallPlanner` owns canonical `host.git.worktree.create` and `host.git.worktree.remove` argument JSON, including branch/base trimming and force flags. `WorkspaceWorktreeOpenEngine` owns the pure local/SSH Remote `Worktree: ...` thread construction, notice payloads, and display labels. The workspace model owns only tool dispatch, local/SSH project registration, selection, persistence, and top-bar refresh after a worktree is created.
 - Workspace status copy and context labels live in `WorkspaceStatusTextBuilder`. `/status`, slash mode confirmations, top-bar subtitles, instruction labels, memory labels, and mode labels must use the same helper so transcript copy, SwiftUI surfaces, static HTML, and menu-bar surfaces do not drift. `WorkspaceModel` owns when to emit a status response; the builder owns what the status response says.
 - Top-bar status semantics live in `QuillCodeTopBarStatusPresentation`, not in SwiftUI or HTML renderers. Agent status tone, indicator visibility, accessibility text, and runtime issue tone are shared presentation values so native UI and static UI harness snapshots stay visually and semantically aligned.
